chore: remove versioning name of carousel
