--- src/components/sections/events/event-card-v2/functions/util/formatDateWithTime.ts
@@ -1,20 +1,0 @@
-export default function formatDateWithTime(
-  startDate: Date,
-  endDate: Date,
-): string {
-  const startDay = startDate.getDate();
-  const startMonth = startDate.toLocaleDateString("en-US", { month: "long" });
-  const startYear = startDate.getFullYear();
-  const startTime = startDate.toLocaleTimeString("en-US", {
-    hour: "numeric",
-    minute: "2-digit",
-    hour12: true,
-  });
-  const endTime = endDate.toLocaleTimeString("en-US", {
-    hour: "numeric",
-    minute: "2-digit",
-    hour12: true,
-  });
-
-  return `${startDay} ${startMonth} ${startYear}, ${startTime} - ${endTime}`;
-}
