+++ src/components/sections/events/event-card/EventCardContent.tsx
@@ -1,10 +1,10 @@
 import { Box } from "@mui/material";
 import EventCardStatus from "./EventCardStatus";
 import EventCardTitle from "./EventCardTitle";
-import { Event } from "@/constants/types/events.type";
-import EventCardTags from "./EventCardTags";
 import EventCardDetails from "./EventCardDetails";
+import EventCardTags from "./EventCardTags";
 import EventCardTicketButton from "./EventCardTicketButton";
+import { Event } from "@/constants/types/events.type";
 
 interface EventCardContentProps {
   event: Event;
@@ -12,7 +12,7 @@ interface EventCardContentProps {
   onGetTicketClick?: () => void;
 }
 
-export default function EventCardContent({
+export default function EventCardConten({
   event,
   showGetTicket,
   onGetTicketClick,
@@ -20,21 +20,33 @@ export default function EventCardContent({
   return (
     <Box
       sx={{
-        padding: "0 16px 16px 16px",
         display: "flex",
         flexDirection: "column",
-        gap: "8px",
-        flex: 1,
+        height: "100%", // Take full height of the 168px container
+        justifyContent: "space-between", // Distribute content and button
       }}
     >
-      <EventCardStatus status={event.status} />
-      <EventCardTitle title={event.title} />
-      <EventCardDetails event={event} />
-      <EventCardTags tags={event.tags} />
-      <EventCardTicketButton
-        showGetTicket={showGetTicket}
-        onGetTicketClick={onGetTicketClick}
-      />
+      {/* Main content at the top */}
+      <Box
+        sx={{
+          display: "flex",
+          flexDirection: "column",
+          gap: "4px", // Same as original
+        }}
+      >
+        <EventCardStatus status={event.status} />
+        <EventCardTitle title={event.title} />
+        <EventCardDetails event={event} />
+        <EventCardTags tags={event.tags} />
+      </Box>
+
+      {/* Button at bottom left */}
+      <Box sx={{ display: "flex", justifyContent: "flex-start" }}>
+        <EventCardTicketButton
+          showGetTicket={showGetTicket}
+          onGetTicketClick={onGetTicketClick}
+        />
+      </Box>
     </Box>
   );
 }
