+++ src/components/sections/events/event-card/EventCardTitle.tsx
@@ -15,6 +15,7 @@ export default function EventCardTitle({ title }: EventCardTitleProps) {
         lineHeight: "109%",
         letterSpacing: "0%",
         textAlign: "left",
+        padding: "4px 0 2px 0",
         background: `linear-gradient(0deg, rgba(240, 240, 240, 0.9), rgba(240, 240, 240, 0.9)),
           linear-gradient(265.86deg, #FFFFFF 29.57%, rgba(236, 236, 236, 0.23) 114.98%)`,
       }}
