+++ src/components/sections/events/event-card/functions/util/getGradientTextStyle.ts
@@ -1,20 +1,15 @@
-import { Poppins } from "next/font/google";
-
-const poppins = Poppins({
-  weight: ["400", "700"],
-  subsets: ["latin"],
-  display: "swap",
-});
-
-export default function getGradientTextStyle(opacity: number = 0.9) {
+/**
+ * Generate gradient text style for consistent text styling across components
+ * @param opacity - The opacity value for the gradient (0-1) - currently not used but kept for compatibility
+ * @returns Style object with gradient text properties
+ */
+export default function getGradientTextStyle(opacity?: number) {
   return {
-    fontFamily: poppins.style.fontFamily,
-    background: `
-      linear-gradient(0deg, rgba(240, 240, 240, ${opacity}), rgba(240, 240, 240, ${opacity})),
-      linear-gradient(265.86deg, #FFFFFF 29.57%, rgba(236, 236, 236, 0.23) 114.98%)
-    `,
+    background: `linear-gradient(0deg, rgba(240, 240, 240, ${opacity}), rgba(240, 240, 240, ${opacity})),
+linear-gradient(265.86deg, #FFFFFF 29.57%, rgba(236, 236, 236, 0.23) 114.98%)`,
+    backgroundClip: "text",
     WebkitBackgroundClip: "text",
     WebkitTextFillColor: "transparent",
-    backgroundClip: "text",
+    color: "transparent", // Fallback for browsers that don't support backgroundClip
   };
 }
