+++ src/components/sections/events/event-carousel/EventCarousel.tsx
@@ -19,11 +19,11 @@ interface EventCarouselProps {
 }
 
 /**
- * EventCarousel Component
+ * EventCarousel Component - Built with EventCard
  *
- * A carousel component that displays event cards in a horizontal scrolling layout.
- * Features navigation buttons on both ends and supports customizable number of events per page.
- * Uses a sliding window approach to smoothly animate between pages of events.
+ * A carousel component that displays EventCard components in a horizontal scrolling layout.
+ * Features glass navigation buttons floating on the sides and proper page-based pagination
+ * to prevent card offset issues.
  *
  * @example
  * // Basic usage with default 3 events per page
@@ -33,20 +33,12 @@ interface EventCarouselProps {
  * // Custom configuration with ticket buttons
  * <EventCarousel
  *   events={eventList}
- *   eventsPerPage={4}
+ *   eventsPerPage={3}
  *   showGetTicket={true}
- *   onGetTicketClick={(event) => window.open(event.registrationLink)}
- *   eventCardWidth={350}
+ *   onGetTicketClick={(event) => console.log(event)}
+ *   eventCardWidth={368}
+ *   eventCardHeight={329}
  *   gap={20}
- * />
- *
- * @example
- * // With custom styling
- * <EventCarousel
- *   events={eventList}
- *   eventsPerPage={2}
- *   containerHeight={500}
- *   sx={{ marginTop: 4 }}
  * />
  */
 export default function EventCarousel({
@@ -54,10 +46,10 @@ export default function EventCarousel({
   eventsPerPage = 3,
   showGetTicket = false,
   onGetTicketClick,
-  eventCardWidth = 320,
-  eventCardHeight = 280,
-  gap = 16,
-  containerHeight = 350,
+  eventCardWidth = 368,
+  eventCardHeight = 329,
+  gap = 20,
+  containerHeight = 400,
   sx = {},
 }: EventCarouselProps) {
   const [currentPage, setCurrentPage] = useState(0);
@@ -69,21 +61,25 @@ export default function EventCarousel({
 
   // Navigation handlers
   const goToPreviousPage = () => {
-    setCurrentPage((prev) => (prev > 0 ? prev - 1 : totalPages - 1));
+    setCurrentPage((prev) => (prev > 0 ? prev - 1 : 0));
   };
 
   const goToNextPage = () => {
-    setCurrentPage((prev) => (prev < totalPages - 1 ? prev + 1 : 0));
+    setCurrentPage((prev) =>
+      prev < totalPages - 1 ? prev + 1 : totalPages - 1,
+    );
   };
 
-  // Calculate container width based on events per page
+  // Calculate exact container width based on events per page (no extra space)
   const containerWidth =
     eventsPerPage * eventCardWidth + (eventsPerPage - 1) * gap;
 
-  // Calculate translate offset for sliding animation
-  const translateX = useMemo(() => {
-    return -(currentPage * containerWidth);
-  }, [currentPage, containerWidth]);
+  // Get events for current page
+  const currentPageEvents = useMemo(() => {
+    const startIndex = currentPage * eventsPerPage;
+    const endIndex = startIndex + eventsPerPage;
+    return events.slice(startIndex, endIndex);
+  }, [events, currentPage, eventsPerPage]);
 
   // If no events, show empty state
   if (events.length === 0) {
@@ -109,116 +105,134 @@ export default function EventCarousel({
     <Box
       sx={{
         display: "flex",
+        flexDirection: "column",
         alignItems: "center",
-        gap: 2,
-        width: "100%",
-        height: containerHeight,
+        height: containerHeight + 40, // Reserve space for page indicators
         ...sx,
       }}
     >
-      {/* Left Navigation Button */}
-      <GlassArrowButton
-        onClick={goToPreviousPage}
-        disabled={totalPages <= 1}
-        direction="left"
-        size={50}
-        sx={{
-          flexShrink: 0,
-          opacity: totalPages <= 1 ? 0.5 : 1,
-          cursor: totalPages <= 1 ? "default" : "pointer",
-        }}
-      />
-      {/* Carousel Container (Overflow Hidden) */}
+      {/* Carousel Container */}
       <Box
         sx={{
-          width: containerWidth,
-          height: "100%",
-          overflow: "hidden",
           position: "relative",
-          flexShrink: 0,
+          width: containerWidth,
+          height: containerHeight,
         }}
       >
-        {/* Sliding Events Container */}
+        {/* Left Glass Button - Floating on the side */}
+        {totalPages > 1 && currentPage > 0 && (
+          <GlassArrowButton
+            onClick={goToPreviousPage}
+            direction="left"
+            size={50}
+            sx={{
+              position: "absolute",
+              left: -25, // Position outside the container, floating above cards
+              top: "50%",
+              marginTop: "-25px", // Half of button size (50px) to center vertically
+              zIndex: 10,
+            }}
+          />
+        )}
+        {/* Right Glass Button - Floating on the side */}
+        {totalPages > 1 && currentPage < totalPages - 1 && (
+          <GlassArrowButton
+            onClick={goToNextPage}
+            direction="right"
+            size={50}
+            sx={{
+              position: "absolute",
+              right: -25, // Position outside the container, floating above cards
+              top: "50%",
+              marginTop: "-25px", // Half of button size (50px) to center vertically
+              zIndex: 10,
+            }}
+          />
+        )}
+        {/* Cards Container - No margins, cards fill the exact width */}
         <Box
           sx={{
             display: "flex",
             gap: `${gap}px`,
+            width: "100%",
             height: "100%",
-            width: `${totalPages * containerWidth}px`,
-            transform: `translateX(${translateX}px)`,
-            transition: "transform 0.3s ease-in-out",
             alignItems: "flex-start",
+            transition: "all 0.3s ease-in-out",
           }}
         >
-          {/* Render all pages */}
-          {Array.from({ length: totalPages }).map((_, pageIndex) => {
-            const pageStartIndex = pageIndex * eventsPerPage;
-            const pageEndIndex = pageStartIndex + eventsPerPage;
-            const pageEvents = events.slice(pageStartIndex, pageEndIndex);
-
-            return (
+          {/* Render only current page events */}
+          {currentPageEvents.map((event) => (
+            <Box
+              key={event.id}
+              sx={{
+                width: eventCardWidth,
+                height: eventCardHeight,
+                flexShrink: 0,
+              }}
+            >
+              <EventCard
+                event={event}
+                showGetTicket={showGetTicket}
+                onGetTicketClick={() => onGetTicketClick?.(event)}
+                width={eventCardWidth}
+                height={eventCardHeight}
+              />
+            </Box>
+          ))}
+          {/* Fill remaining slots if current page has fewer events than eventsPerPage */}
+          {currentPageEvents.length < eventsPerPage &&
+            Array.from({
+              length: eventsPerPage - currentPageEvents.length,
+            }).map((_, index) => (
               <Box
-                key={pageIndex}
+                key={`empty-${index}`}
                 sx={{
-                  display: "flex",
-                  gap: `${gap}px`,
-                  width: containerWidth,
-                  height: "100%",
+                  width: eventCardWidth,
+                  height: eventCardHeight,
                   flexShrink: 0,
-                  alignItems: "flex-start",
+                  visibility: "hidden", // Invisible placeholder to maintain layout
                 }}
-              >
-                {/* Render events for this page */}
-                {pageEvents.map((event) => (
-                  <Box
-                    key={event.id}
-                    sx={{
-                      width: eventCardWidth,
-                      height: eventCardHeight,
-                      flexShrink: 0,
-                    }}
-                  >
-                    <EventCard
-                      event={event}
-                      showGetTicket={showGetTicket}
-                      onGetTicketClick={() => onGetTicketClick?.(event)}
-                      width={eventCardWidth}
-                      height={eventCardHeight}
-                    />
-                  </Box>
-                ))}
-
-                {/* Fill remaining slots with empty spaces to maintain left alignment */}
-                {Array.from({ length: eventsPerPage - pageEvents.length }).map(
-                  (_, emptyIndex) => (
-                    <Box
-                      key={`empty-${pageIndex}-${emptyIndex}`}
-                      sx={{
-                        width: eventCardWidth,
-                        height: eventCardHeight,
-                        flexShrink: 0,
-                        visibility: "hidden",
-                      }}
-                    />
-                  ),
-                )}
-              </Box>
-            );
-          })}
+              />
+            ))}
         </Box>
       </Box>
-      {/* Right Navigation Button */}
-      <GlassArrowButton
-        onClick={goToNextPage}
-        disabled={totalPages <= 1}
-        direction="right"
-        size={50}
+
+      {/* Page Indicator Dots - Fixed position with reserved space */}
+      <Box
         sx={{
-          flexShrink: 0,
-          opacity: totalPages <= 1 ? 0.5 : 1,
-          cursor: totalPages <= 1 ? "default" : "pointer",
+          display: "flex",
+          gap: 1,
+          justifyContent: "center",
+          alignItems: "center",
+          height: "32px", // Fixed height for page indicators area
+          marginTop: "20px",
         }}
-      />
+      >
+        {totalPages > 1 &&
+          Array.from({ length: totalPages }).map((_, index) => (
+            <Box
+              key={index}
+              onClick={() => setCurrentPage(index)}
+              sx={{
+                width: index === currentPage ? 12 : 8,
+                height: index === currentPage ? 12 : 8,
+                borderRadius: "50%",
+                backgroundColor:
+                  index === currentPage
+                    ? "#ffffff"
+                    : "rgba(255, 255, 255, 0.4)",
+                cursor: "pointer",
+                transition: "all 0.3s ease",
+                ...(index !== currentPage && {
+                  "&:hover": {
+                    backgroundColor: "#ffffff",
+                    transform: "scale(1.4)",
+                  },
+                }),
+              }}
+            />
+          ))}
+      </Box>
     </Box>
   );
 }
